pyproject.toml
[project]
name = "arclet-alconna-avilla"
version = "0.15.0"
description = "Support Alconna for GraiaProject/Avilla"
authors = [
    {name = "RF-Tar-Railt", email = "rf_tar_railt@qq.com"},
]
dependencies = [
    "arclet-alconna<2.0.0, >=1.7.14",
    "arclet-alconna-tools<0.7.0, >=0.6.3",
    "tarina>=0.3.3",
    "nepattern<0.6.0, >=0.5.10",
    "creart-graia>=0.1.5",
    "creart>=0.2.1",
    "graia-amnesia>=0.5.0",
    "graia-broadcast>=0.18.2",
    "graia-saya>=0.0.17",
    # "avilla",
    # "launart"
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "AGPL-3.0"}

[tool]
[tool.pdm]
[tool.pdm.build]
includes = ["arclet/alconna/avilla"]
[tool.pdm.dev-dependencies]
dev = [
    "nonechat>=0.2.0",
]
